+++ GFramework.Core.Tests/Coroutine/QueryCoroutineExtensionsTests.cs
@@ -329,63 +329,3 @@ public class QueryCoroutineExtensionsTests
         Assert.That(receivedResult, Is.Null);
     }
 }
-
-/// <summary>
-///     用于测试的整数查询类
-/// </summary>
-internal class IntQuery : IQuery<bool>
-{
-    private IArchitectureContext? _context;
-    public int Value { get; set; }
-
-    public void SetContext(IArchitectureContext context)
-    {
-        _context = context;
-    }
-
-    public IArchitectureContext GetContext()
-    {
-        return _context!;
-    }
-
-    public bool Do()
-    {
-        return Value > 0;
-    }
-}
-
-/// <summary>
-///     用于测试的复杂查询类
-/// </summary>
-internal class ComplexQuery : IQuery<ComplexResult>
-{
-    private IArchitectureContext? _context;
-    public string Name { get; set; } = string.Empty;
-    public List<int> Values { get; set; } = new();
-    public Dictionary<string, object> Metadata { get; set; } = new(StringComparer.Ordinal);
-
-    public void SetContext(IArchitectureContext context)
-    {
-        _context = context;
-    }
-
-    public IArchitectureContext GetContext()
-    {
-        return _context!;
-    }
-
-    public ComplexResult Do()
-    {
-        return new ComplexResult { ProcessedName = Name, Sum = Values.Sum(), Count = Values.Count };
-    }
-}
-
-/// <summary>
-///     用于测试的复杂结果类
-/// </summary>
-internal class ComplexResult
-{
-    public string ProcessedName { get; set; } = string.Empty;
-    public int Sum { get; set; }
-    public int Count { get; set; }
-}
